requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.1.8
feedparser==6.0.11
Flask==3.1.0
Flask-SocketIO==5.5.1
Flask-SSE==1.0.0
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector==2.2.9
numpy==2.2.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
redis==5.2.1
regex==2024.11.6
setuptools==75.8.0
sgmllib3k==1.0.0
simple-websocket==1.1.0
six==1.17.0
tzdata==2025.1
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
